refactor: extract isOwnOidcCallback + tighten handleLogin return
