+++ src/state/renewingAtom.ts
@@ -1,6 +1,6 @@
 import { atom } from 'jotai';
 
-// True while any silent-renew is in flight; consumers gate recovery paths on it
-// so a transient 401 during renewal isn't misread as a session drop.
+// True while a silent renew is in flight. Recovery paths check this so
+// a transient 401 during renewal isn't treated as a session drop.
 export const renewingAtom = atom<boolean>(false);
 renewingAtom.debugLabel = 'renewingAtom';
